requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
SQLAlchemy==2.0.35 
aiomysql==0.2.0
PyMySQL==1.1.1 
cryptography==43.0.1
passlib[bcrypt]==1.7.4
pydantic>=2,<3
email-validator>=2.0.0
python-jose[cryptography]==3.3.0
python-multipart
boto3==1.35.21